Conflicts:
	applications/kiptool/copy_dlls.bat
	applications/kiptool/qt/QtKipTool/QtKipTool.pro
	applications/kiptool/src/kiptoolmainwindow.cpp
	applications/kiptool/src/kiptoolmainwindow.h
	core/kipl/kipl/include/io/io_tiff.h
	core/kipl/kipl/include/math/findpeaks.h
	frameworks/imageprocessing/ProcessFramework/include/KiplEngine.h
	frameworks/imageprocessing/ProcessFramework/include/KiplModuleItem.h
	frameworks/imageprocessing/ProcessFramework/include/KiplProcessModuleBase.h
	frameworks/imageprocessing/ProcessFramework/src/KiplEngine.cpp
	frameworks/imageprocessing/modules/BaseModules/qt/BaseModules/BaseModules.pro
	frameworks/imageprocessing/modules/ClassificationModules/qt/ClassificationModules/ClassificationModules.pro
	frameworks/imageprocessing/modules/ImagingModules/qt/ImagingModules/ImagingModules.pro

+++ applications/kiptool/src/fileconversiondialog.cpp
@@ -119,7 +119,7 @@ void FileConversionDialog::on_pushButton_StartConversion_clicked()
 
     msg.str("");
     msg<<"File list size="<<plist.size();
-    ui->progressBar->setMaximum(flist.size());
+    ui->progressBar->setMaximum(static_cast<int>(flist.size()));
     ui->progressBar->setValue(0);
     logger(logger.LogMessage,msg.str());
